third_party/blink/web_tests/dom/mutation-event-tests/fast/events/mutation-during-replace-child-expected.txt

CONSOLE WARNING: Listener added for a 'DOMNodeRemoved' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
Ensures that replaceChild() throws an exception if mutation even handler does something wrong

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS target.replaceChild(newChild, oldChild); threw exception NotFoundError: Failed to execute 'replaceChild' on 'Node': The node to be removed is no longer a child of this node. Perhaps it was moved in response to a mutation?.
PASS successfullyParsed is true

TEST COMPLETE
